Consolidate env_vars and file_ingredients helper functions into utils.
